backend/requirements.txt
pathway>=0.13.0
fastapi>=0.110.0
uvicorn>=0.29.0
openai>=1.14.0
reportlab>=4.1.0
python-dotenv>=1.0.1
requests>=2.31.0
pydantic>=2.0.0
aiofiles>=23.0.0
httpx>=0.27.0
